+++ netbox_zabbix/jobs/__init__.py
@@ -1,0 +1,6 @@
+from .zabbix import *
+
+__all__ = (
+    'update_zabbix_device',
+    'update_zabbix_vm',
+)
