+++ netbox_zabbix/signals/__init__.py
@@ -1,0 +1,7 @@
+from .signals import *
+
+__all__ = (
+    'update_device',
+    'update_vm',
+    'm2m_device',
+)
